order_chat_bot/requirements.txt
flask==2.3.3
twilio==8.10.0
openpyxl==3.1.2
pandas==2.1.1
python-dotenv==1.0.0
nltk==3.8.1
openai>=1.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
firebase-admin>=6.0.0
requests>=2.25.0